% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronecker.R
\name{fast_kronecker}
\alias{fast_kronecker}
\alias{fast_kronecker,dgCMatrix,dgCMatrix-method}
\alias{fast_kronecker,dsCMatrix,dsCMatrix-method}
\alias{fast_kronecker,dsCMatrix,ddiMatrix-method}
\title{Faster kronecker product for sparse matrices}
\usage{
fast_kronecker(a, b)

\S4method{fast_kronecker}{dgCMatrix,dgCMatrix}(a, b)

\S4method{fast_kronecker}{dsCMatrix,dsCMatrix}(a, b)

\S4method{fast_kronecker}{dsCMatrix,ddiMatrix}(a, b)
}
\arguments{
\item{a}{Left operand, can be dgCMatrix or dsCMatrix}

\item{b}{Right operand, must have same type as left operand}
}
\description{
A much faster sparse kronecker product than that in the Matrix package, with
methods for symmetric and general matrices. Currently this method doesn't
work for all combinations of inputs, so check the parameter type
documentation.
}
\section{Functions}{
\itemize{
\item \code{fast_kronecker,dgCMatrix,dgCMatrix-method}: General and general.

\item \code{fast_kronecker,dsCMatrix,dsCMatrix-method}: Symmetric and symmetric.

\item \code{fast_kronecker,dsCMatrix,ddiMatrix-method}: Symmetric and diagonal
}}

