\name{FastSparse-package}
\alias{FastSparse-package}
\alias{FastSparse}
\docType{package}
\title{
\packageTitle{FastSparse}
}
\description{
\packageDescription{FastSparse}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FastSparse}
\packageIndices{FastSparse}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FastSparse}

Maintainer: \packageMaintainer{FastSparse}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
