% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.FastSparse}
\alias{print.FastSparse}
\alias{print.FastSparseCV}
\title{Print FastSparse.fit object}
\usage{
\method{print}{FastSparse}(x, ...)

\method{print}{FastSparseCV}(x, ...)
}
\arguments{
\item{x}{The output of FastSparse.fit or FastSparse.cvfit}

\item{...}{ignore}
}
\description{
Prints a summary of FastSparse.fit
}
